pyproject.toml
[build-system]
# grpcio-tools is required at build time to regenerate the protobuf stubs
# from proto/gateway.proto via hatch_build.py.
requires = ["hatchling", "grpcio-tools"]
build-backend = "hatchling.build"

[project]
name = "ros_core_client"
version = "0.2.0"
description = "ROS core client — ROS1 and ROS2 environment and communicator"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Yongshuai Wu", email = "ywu26@students.kennesaw.edu" }
]
license = { file = "LICENSE" }
dependencies = [
    "grpcio",
    "numpy",
    "protobuf"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]

[project.urls]
Homepage = "https://jlab-git.ayayadomain.com/gitlab/ayaya/ros_core_client.git"
Repository = "https://jlab-git.ayayadomain.com/gitlab/ayaya/ros_core_client.git"
Documentation = "https://jlab-git.ayayadomain.com/gitlab/ayaya/ros_core_client.git#readme"

# ── Hatchling build hook ──────────────────────────────────────────────────────
# hatch_build.py (repo root) runs grpc_tools.protoc to regenerate
# ros_core_client/ros2/proto/gateway_pb2.py and gateway_pb2_grpc.py from
# proto/gateway.proto, then patches the absolute import in gateway_pb2_grpc.py
# to a relative import so the stubs work correctly inside the package.
[tool.hatch.build.hooks.custom]
path = "hatch_build.py"

# Ensure the proto source directory and generated stubs are included in sdist.
[tool.hatch.build.targets.sdist]
include = [
    "proto/",
    "ros_core_client/",
    "hatch_build.py",
    "pyproject.toml",
    "README.md",
    "LICENSE",
]
